official: Upgrade container to CentOS 6
CentOS 5 is not supportable anymore due to its insecure OpenSSL
implementation, which makes it impossible to fetch software from certain
web sites.  We needed to upgrade the assembler in order to build the
Huffman encoding optimizations in the evolving libjpeg-turbo 2.1 code
base, and this was the only way to accomplish that.